updating tests for the various versionsof mth5 files
