+++ tests/version_2/__init__.py
@@ -1,0 +1,1 @@
+"""Unit test package for mth5."""
