+++ tests/version_2/test_from_stationxml.py
